Added warning suppression for unused variable
When in release mode, which adds optimization to strip out assert calls,
the variable assigned to the return code of the transformation call is
then not used (only used in comparison within assert call, which got
stripped). Added a trivial use of the variable after assert to suppress
warning in release mode.