Added support for EoL comments, and support for Boost 1.46.1
There is an overhaul in boost::spirit::qi between 1.46.1 and 1.47.0,
making the previous version of the code not compile against Boost
1.46.1. The structure of the grammar has been modified such that it
compiles against versions 1.46.1 and up. This meant that config_block
values are initially read in with trailing whitespace included, but the
trailing whitespace is trimmed during a post-processing step.