fix a failing test on descriptor_set conversions
Previously this test was using matrices with uninitialized values
to test the conversion of ocv descriptors.  The uninitialized values
were treated as random or arbitrary.  Unfortunately, on some platforms,
some of these uninitialized value where "nan".  Since (nan == nan) is
false, the test indicated that conversion was failing.  This fix
initializes the input data to random value generated with cv::randu(),
so there should be no nan or infinite values.