Addressed static building separating APM + baking-in module
In order to stabilly build static libraries, we disable the dynamic
library loading that would normally occur in the
algorithm_plugin_manager in favor of linking all active plugin libraries
into the, now separate, algorithm plugin management (APM) library. This
occurrs only when building statically. Building shared (dynamically)
behaves as it did before, allowing plugin library pick-up at runtime.
The only impact to shared mode is the existance of the new APM library
that must be linked to for the plugin manager to be used.