Added camera_map interface; added reference counting to interface caches
Interface cache had an issue where, if the same underlying object was
extracted and "stored" in the cace multiple times, it only took the
erasure of one of those to invalidate that instance in the C interface
(cache object would throw as it wouldn't see the instance's ptr in the
map post erasure). So, I added store/erase count balancing, i.e. erase
will only actually erase the ptr from the cache when erase is called the
same number of times as store for a given pointer.