Redefined algorithm registration to allow overwriting
Before, the registration function, given a label-instance pair, would
return false if a given label for a type was registered a second time.
This, however, would prevent the ability to reload plugins during
runtime or other complex behaviours, as well as leading to confusing
error/warning messages saying that particular plugins failed to register
when they were really just duplicately registered (not an error).

- Refined registrar use of typedefs
- Added check method for registered labels for a type