Fix plugin build and install paths on Windows
Plugins are built as "modules" which means that they are built in
the library path (/lib) instead of the runtime path (/bin).  Other
shared libraries (DLLs) in Windows are built in the runtime path.
As a result, no distinction between Windows and other platforms
is needed regarding where to find plugins.