Add a "_plugin" suffix to the base name for plugins
It was found to be problematic, especially on Windows, for the base
library and plugin to have the same name but be in different paths.
The Windows LoadLibrary call would load the plugin, then search for
the base library of the same name and find the plugin again because
the first search path was the plugin directory.