Fixed a bug in generating noisy landmarks for tests.
This bug was not causing any problems yet, but could in the future.
The function was modifying the landmarks in place as well as putting
them into a new landmark map, so the original data was modified.
This would be a problem for tests that want to compare the results
of optimization to the original noise-free values.  This change
clones each landmark before modification to preserve the original data.