Add display of cameras
Implement a rudimentary and incomplete visualization of camera frustums.
This is incomplete as we do not have sufficient information to compute
the frustum FOV angle or aspect ratio. (Also, the size is effectively
arbitrary, although this might be a feature.)

This is likely not sufficient in the long term, as it doesn't readily
provide a way to highlight individual cameras.