Refactor camera management
Create data structures to hold information about a camera and associated
image. Refactor how we load cameras to make use of this, and associated
new helper functions. Also, when loading a camera without an image,
"guess" the image size by assuming that the principle point is exactly
the image center, which allows us to compute more reasonable values for
the camera FOV and aspect ratio.

This is in preparation for actually keeping track of this information in
order to be able to do things like have a "list" of cameras and identify
which is active, and, eventually, provide a "camera view" presentation
of the data in addition to the "world view".