Allow loading camera from image path and directory
Add an overload of read_krtd_file that takes an image file path and
camera directory, like read_krtd_files, but for reading only one camera.
Modify read_krtd_files to use this.

This will be useful to the GUI that wants to traverse the list itself
(partly because it needs to do so anyway, and partly to avoid gratuitous
type conversions).