Make camera view work without images
Modify CameraView to cache the image bounds, and to fabricate bounds
from an image size if no image is given. Modify MainWindow to pass its
notion of the image dimensions to CameraView::loadImage.

This allows us to provide a reasonable visualization if only landmarks
and cameras (but no frame images) are loaded.