Revert change to camera loading logic
Revert the change to MainWindowPrivate::addCamera to take a path rather
than a camera object that moved the loading into said method. This
restores the previous behavior of using the maptk library logic to
determine the camera KRTD file path when loading a project, so that we
are not duplicating the logic used by maptk::read_krtd_files.

Also, move more of the logic to fill the CameraData struct into
MainWindowPrivate::addCamera, since most of it had been moved there
already anyway.