Fix logic to update image dimensions
Fix the logic that updates the image dimensions on the vtkMkCamera
object when loading a camera frame image, by moving it into
CameraView::loadImage. The logic from 8837890429d9 did not actually
propagate the updated dimensions, and so effectively did nothing.
Instead, pass the camera instance to CameraView::loadImage and have that
update the camera's image dimensions when an image is successfully
loaded.