Don't leak implementation details
Remove updateCameraActors method from WorldView. The way this was being
used forces users to be aware of implementation details, which is never
desirable, and was incorrect because loading cameras individually was
not calling it. For now, we update any time a change happens, which is
inefficient. We could add update coalescing to WorldView (and there is a
TODO to do exactly that), but an even better solution would be to push
the logic all the way back into vtkMaptkCameraRepresentation (which has
the same 'leak' problem).