Fix more update logic
Remove public method to recompute camera scales; as in b367a4187bc3,
this is leaking implementation details and should happen automatically
as needed. Also implement logic to squash multiple updates, so that we
aren't penalized by recomputing for e.g. every camera when adding a
whole bunch of cameras at once (thus addressing the efficiency comment
from the aforementioned commit).

Setting the initial camera scale has also been removed; it is
superfluous with what is done in the representation's constructor.
Eventually this should be made user configurable, but that will involve
interaction with a helper class that would result in this code needing
to be removed anyway; doing it now ensures we won't forget later.