Start implementing UI for CameraView
Refactor WorldView to have a .ui file, which allows us to add a tool bar
to the widget that can provide view-specific actions. Add action to
reset the view (also to full extents) and modify automatic reset logic
to only reset on initial frame, rather than frame changes. (Now that we
can do a manual reset, we don't have a need to have it happen some other
way, and resetting on frame changes was otherwise undesirable.)