"Fix" camera scaling
Change base scale factor from 10% to 20%.  Re-add scale change to
inactive cameras that was accidentally dropped in 1e83be25f97e.  Fix
incorrect calculation of scale due to the camera representation not
being updated. (We now cause a camera update to also trigger a scale
update, as well as update the representation when computing the scale.
This is a bit redundant and verging on paranoia, but should prevent
accidental errors, and the way we use the dirty flags should ensure we
don't ever actually do more work than needed.)