Hook up (most) camera display options
Implement and hook up most of the camera visualization options. This
includes both pushing the initial persisted state to the visualization,
and updating it when the user makes changes.

Display of inactive cameras as points is not currently implemented in
the representation, so the corresponding UI controls are not currently
functional.