Refactor vtkMaptkCameraRepresentation
Add a PIMPL class and move members that don't need to be exposed to the
same; this allows us to convert several pointers (some smart, some bare)
into vtkNew. Make BuildFrustum a private non-member function, as it does
not need to modify any members. Use explicit deletion on copy ctor and
assignment operator, as we require that feature (via qtExtensions)
anyway, and it produces better error messages. Add a private helper
method to get the next camera (makes those loops prettier). Keep dummy
vtkPolyData around as a private member, as we need to re-add it to the
vtkAppendPolyData when resetting the inactive cameras to avoid VTK
complaining (a regression that was previously fixed in 3c4acc4b6fa8; see
the same for more details about the problem being fixed).