Disable lighting for active camera
Turn off lighting for the active camera actor. Applying lighting to
wireframes is somewhat atypical in general, and while it can make some
sense for the inactive cameras (which may be dense enough for the
lighting to help visualize the 3D shape of the camera collection), it
doesn't make as much sense for the active camera, and often just results
in the camera being hard to see against the black background.