Add options for world view landmark display
Create another options class to allow modifying some (for now) basic
properties of point clouds, i.e. point size and color. Hook this up to
allow changing the landmark visualization in the world view. (We'll be
doing this for the feature points and landmarks in the camera view also,
but in a separate commit.)

Also, use QList instead of vtkActorCollection to keep track of the
landmark actors. This is easier to use (at least for iteration) and
lighter weight; we don't need the extra features of the latter.