Create manual for GUI
Import user manual build system from ViViA. Create an initial version of
a user manual for the GUI (generally "complete" for the current state of
the GUI). Create 16x16 versions of icons, as these work better in the
manual.

For now, the manual is only accessible offline; accessing the manual
directly from within the GUI will be added at a later time.