Add "application data"
Create build machinery to generate detailed build information, which
will be displayed in (for now, just) the GUI. Add an ACKNOWLEDGMENTS
file that will also be used in the GUI. Add a Qt resource script to
make these (and also LICENSE) available to the GUI as Qt resources.

All of this will eventually be used in an About dialog.