Rename FeatureOptions -> PointOptions
Rename the visualization options pop-up per summary. The original intent
was for "feature" in this context to imply a generic name including both
feature points and landmarks, but we would like to extent the dialog for
feature points to include options for feature track trails. The rename
frees up the "FeatureOptions" name for this purpose, and anyway,
"Points" works better for the generic version.