Improve vtkMaptkCameraRepresentation performance
Refactor update logic in vtkMaptkCameraRepresentation so that we always
keep a copy of the polydata for non-active cameras (allowing us to reuse
these, rather than rebuilding the polydata for every non-active camera
to be rendered every update). Also, don't update the path or active
camera except when actually needed.

This is effectively a reworking of a patch to this effect originally
written by Rusty.