Remove apparently spurious self-dependency
Remove doxygen tag dependency of "maptk" documentation on the "MAP-Tk"
tags. There doesn't seem to be anything that would ever create this, and
it results in a broken build graph when (API) documentation is enabled.
(This seems to be a new development since switching to Vital. Not clear
why it was working before.)

At any rate, it seems suspicious that the documentation depends on "its
own" tags, and more likely that the specified dependency is an accident.