Add support for missing cameras
Modify GUI to retain camera data entries when a camera cannot be loaded.
Modify world view to accept a null camera (and track when this means
there is no camera transform as a result in order to hide the projected
image). This fixes some glitches when cameras were missing due to a
disconnect in camera data indices and camera ID's, which are expected to
be the same, and allows us to display feature points and images in the
camera view when the camera model is missing.