Use foreach / VITAL_FOREACH in GUI
Convert some loops over container values that were superfluously using
foreach_iter (or worse, the equivalent, set up longhand) to use foreach
or VITAL_FOREACH instead. The Qt version of the former makes an
unnecessary copy of the container, but the QtExtensions version (which
replaces it) uses a C++11 range-based for, or a compatible equivalent
where range-based for is not available. (VITAL_FOREACH - used in the VTK
extension classes that contain no Qt code - is essentially the same, but
with an additional option to use BOOST_FOREACH if that is available and
range-based for is not.)