Support loading images directly
Add support for direct image loading to the GUI. This "works", however
both loading images and loading cameras create new "frames", so there is
currently no way to load both the image and camera for a frame (other
than using a project file). This will probably need to be addressed at
some point.