Combine separately loaded images and cameras
Implement logic to combine images and cameras when loaded separately.
This uses fairly strict FIFO queuing and does not allow for "gaps" of
frames having only an image or only a camera in the middle of a
sequence, nor does it attempt in any way to perform matching based on
the file names of the images/cameras being loaded.