More robust "project" support
Modify GUI "project" support to be more robust against values in the
configuration file that are missing or have different names, and to load
what data is available in case some things are missing. In particular,
it is now possible to use the tracking configuration file to load images
and tracks.