Add glossary to user manual
Add a brief glossary section with short definitions of several of the
more important "technical" terms used in the GUI (and MAP-Tk in
general). This is especially important for "track", which has a specific
meaning in Structure from Motion research that does not exactly match up
with what a casual user might expect. (This also "solves" some of our
ongoing terminology headaches by essentially punting on trying to use
"casual user" phrasing that is inconsistent with our internals.)