Switch to new Travis CI infrastructure
This patch changes the way dependencies are built for MAP-Tk.
We now use the new container-based infrastructure which is faster
but limits packages with can install with apt.  Instead, many packages
are built from source and then cached for future builds.

At this point the CI tests are only building a subset of the MAP-Tk components.