Add and install config for bundle adjustment
Add bundle_adjust.conf supplied by Matt Leotta for use as the default
configuration for running bundle adjustment within the GUI. Set up rules
to copy this to the build directory in a location that should be picked
up by the config search when running from a build directory. Also add
rules to install it to the usual FHS-conforming location. (Possibly
we'll want to refine this later to use more "usual" locations for
Windows and OS/X, but this will do for now.)