Allow changing match matrix Y axis direction
Modify the logic dealing with coordinates in the match matrix image to
remove the explicit Y flip to produce graph (Y-up) coordinates rather
than image (Y-down) coordinates, and instead add a new option
("orientation", with the old orientation now "layout") to choose between
these coordinate systems. The new option uses a scene transformation,
allowing us to make the switch without recomputing the image, and
leverages item local coordinates so that we don't need to worry about it
for showing values in the status bar.