More documentation tweaks
Update 'reverse' documentation (both manual and tool tip) to reflect the
state of the new and improved version of the algorithm (which is in a
different branch, but has been merged as of the time these changes are
being made). Remove restriction that 'reverse' must have cameras (again,
to account for upstream changes to the algorithm). Improve documentation
for 'align' (a.k.a. 'canonical transform').