Add landmark colors to camera view
Modify camera view to accept landmark data in a single dump (rather than
resetting it when the active frame changes), which allows us to keep
track of things like whether or not we have meaningful colors. Modify
internal logic when adding a point to look up and use the corresponding
data. Make necessary changes to VTK objects to display colors for
landmarks.