Fix bug when loading (only) images
Fix a potential crash (due to a VTK assert) when updating the camera
scale when we don't actually have any cameras or landmarks loaded. (This
can happen when we load images with no other data loaded, as we can
trigger changes of the active camera - which triggers a scale update -
in such an instance, but don't have any actual geometry from which to
compute a scale. vtkBoundingBox doesn't like being asked for its
diagonal length in such cases.)