Start adding data field visualization support
Add additional logic to PointOptions to allow selecting a data field to
be used for visualization. Separate assignment of actors and mappers;
there is no reason they need to be coupled, and in one case, this is
necessary as we want to add the actor before the mapper has been fully
set up, but we need the mapper set up as soon as it is added in order to
set the initial display state.

Note that this doesn't yet set the color mapping.