Create a scalars to colors helper class
Create a subclass of vtkScalarsToColors that implements mapping of
scalars to colors by direct use of a qtGradient. This avoids us having
to muck about with VTK look-up tables in order to provide the user with
configurable color mapping, and instead allows us to more directly make
use of our recently added GradientSelector.

This is just the class itself. Actually connecting it to the plumbing
and making use of it will come later.