Refactor color-by-data UI
Create a new UI for providing color-by-data options, including the
ability to override the automatic display range. Move data field
selection in the UI so that it is no longer tightly coupled to coloring
by data (which we'll want when we add filtering support). Modify
PointOptions to apply scalar-data coloring, including the correct
scalars-to-colors instance, when selected.