Remove margin from feature options UI
Remove the margin from the FeatureOptions widget. Because this is
embedded in a PointOptions, we were ending up with double margins.
Besides being superfluous, this caused the feature-specific UI to not
line up with the general point UI. (This used to not be terribly
noticeable, but the new check box for data filtering made it somewhat
glaring.)