Set color on track feature points
Add logic to track_features to reconstruct the track set after each
iteration in order to assign a color value to each (new) feature. The
color value used is color at the pixel in the original input image
(before any conversions are applied) that corresponds to the feature
image location.

This is not exactly the most efficient thing to be doing, but currently
there is no way to modify the tracks or features in place, making it
necessary to reconstruct them.

Modify