Compute landmark colors
Add code to bundle_adjust_tracks to compute the landmark colors by
taking the average of all associated feature points. There are some
flaws with this (e.g. it should probably ignore feature points that were
not used to compute the landmark), and probably it really ought to be an
algorithm, but for a first pass, it will do.