Support newer version of VXL where some variables have changed.
The lastest VXL master uses VXL_CORE_INCLUDE_DIRS instead of
VXL_CORE_INCLUDE_DIR to specify multiple include directories (e.g. in source
and build trees).  This is needed to find configured headers that are in
the build tree.  The same is true for VXL_VCL_INCLUDE_DIRS.  This patch
uses the newer variables if available.