Style tweaks; add clang-format options
Apply some style changes (some of which fix definite violations, some of
which were tolerable but now "more consistent") as suggested by
clang-format, and add the style rules file used. This does NOT blindly
accept everything that clang-format wants to do, some of which is
contrary to our accepted style.