Fix include guards
Fix header multiple-inclusion guard symbols of recently added files to
conform to the naming convention in use by the rest of the GUI code
(which is almost, but not quite, the same as the overall MAP-Tk
convention). Remove repetition of guard symbols next to the `#endif`;
they aren't necessary, and this is again consistent with the rest of the
GUI code.