Assorted trivial cleanup
Remove gratuitous section markers; these are useful when the marked
sections span significant portions of the source file, but not so much
when the section contains only a single definition and/or at most a
dozen or so lines of code. Reorder d-function implementation macros
(which should immediately follow includes in most cases) and methods of
private classes (which should precede members). Remove UI state
instances where nothing is being persisted. Remove superfluous empty
constructors of private classes.