Load depth map file list as part of Project
Move logic to read the depth map image file list from MainWindow to
Project, for consistency with the camera image file list. The
refactoring to make use of this also includes a fix for an unchecked
array bound, and changes the name of the camera data member to be
consistent with use elsewhere.