Fix Travis CI cached builds by directly installing Fletch
The previous configuration built Fletch packages and created a local
"install" directory.  This was then copied to the cached install
directory.  This worked for most packages, but VXL includes the absolute
path to some installed subdirectories in its CMake config.  This makes
the VXL install non-relocatable.  The side effect of this was that
initial Travis-CI builds succeeded but builds using a cache of the
Fletch build failed.

The fix provided here is to configure Fletch to install directly to the
cached directory and make the install step in the script a no-op.