Fix camera path when cameras are not added in order
Previously the camera path was drawn as a polyline connecting cameras in
the order they were added to the GUI.  They should be drawn in frame id
order regardless of what order they are added in.  This change stores
the cameras in a sorted id/camera map so they are drawn correctly.  It
also changes some of the camera APIs to focus on frame id rather than
camera pointer.