Make the script more memory effecient
Previously the script would load all images into memory in advance to
get the image sizes for generating the output image bounds.  This is not
practical for large data sets.  Now we read all the headers first to get
the image sizes first.  Then the pixel of image are loaded, used, and
discarded in the main loop.