Allow GUI to load configs with video_source
The new video-based tools use config files with "video_source" instead
of "image_list_file".  When the source is an image list it still refers
to the same file list.  This changes lets the GUI load images from
"video_source" instead of "image_list_file" if that is what is
available.