Fixed display of frame numbers in match matrix window status bar
The matrix indices were being displayed as frame numbers in the Match
Matrix Window status bar.  This was correct if the match matrix
contained every frame starting from zero, the most common case.
However, it was wrong when the video does not start at the beginning or
has skipped frames.  This commit uses the frame number vector to look up
the appropriate frame number for display.