Removed residual testing files
All the unit tests from MAP-Tk have moved into KWIVER along with the
code that was being tested.  For some reason some of that code was left
behind here and was just testing the KWIVER code in the same way that
KWIVER does.  This commit removes all the tests from MAP-Tk, but leaves
the testing directory as a stub.  A future branch should add tests for
the few functions that remain in MAP-Tk and tests that run the command
line tools.