Use CMAKE_SHARED_MODULE_SUFFIX for KWIVER plugins.
KWIVER plugins are built as modules using CMAKE_SHARED_MODULE_SUFFIX for
the file extention.  CMAKE_SHARED_LIBRARY_SUFFIX is similar on most
platforms but not quite correct because MacOS uses .dylib for dynamic
libraries but .so for runtime loadable modules.  As a result the
installer package on MacOS was missing all of the KWIVER plugins and
their dependencies.