Clear the active camera when switching to an invalid frame.
The code was missing a call to reset the active camera to null when changing
to a frame that did not contain a valid camera.  The result of this bug was
that the previous valid camera was used to render a frame with no valid camera.
Instead, the camera in the worldview should disappear when no data is available.