Project now saves to config file.
When new project is created it is saved to a config file. If a
video is opened it is added to the project. If a video is opened
after a project is created it is also added to the project. Tested
the created project files and they correctly reopen in TeleSculptor.