Partial fix to the frame selection problem for compute depth.
This change enforces that the reference frame is the current frame in the
GUI and returns if the refrence frame does not have a valid camera or image.
It also grabs the next "halfsupport" number of valid frames following the
reference frame.  It does not try get "halfsupport" more frame before the
reference frame.