Implement metadata display
Add method to metadata view to receive metadata for a frame and update
the UI with the metadata values. Modify main UI to call this with the
appropriate metadata when changing frames.

For now, we are just relying on metadata_item::as_string, which is
rather ugly in some cases. We may want to add special handling for some
specific types of metadata in the future.