Use QScopedPointer to store project
Rename `currProject` to `project` to avoid an awkward abbreviation. Use
QScopedPointer to store the current project instead of std::shared_ptr;
QScopedPointer fits more naturally with the surrounding Qt code, and
there is no reason to add the overhead of a shared pointer for something
that isn't ever actually shared. Modify project loading to use a
temporary project for parsing, so that if the load fails, the previously
loaded project (if any) will still be loaded.