Open specific file types
Refactor how the user asks to open a file to require that the user
specifies the type of file to be opened, by having separate actions or
CLI options for each file type. This avoids ambiguities that would occur
for example if we want to support .txt as an image list (as our feature
tracks also use .txt).