Use namespace alias in more places
We use kwiver::vital a lot; b19af8e34376 found this sufficiently
egregious to add a short alias for this namespace to save line length,
but for the sake of minimizing code churn at the time, punted on
applying it consistently. Do that now (in MainWindow.cxx), and also use
it in MatchMatrixWindow.cxx in order to shorten an overly long line. (A
subsequent commit is going to generally clean those up...)