Fix non-project config management
Keep a "freestanding" config_block in MainWindow for use when there is
no active project. Merge into this (rather than replacing its contents)
whenever imagery or masks are loaded. Use this as the seed when creating
a project, rather than reloading from the individual files from which
this config is pieced together.

Besides just being more correct, this will be important when we start
implementing more complex mechanisms of setting up video sources.