Fixed version CMake variables
The introduction of the superbuild moved the "project(MAPTK)" call
after the calls to set MAP-Tk version numbers.  The project() call
clears those version variables if a version is not passed to it.
This change resolves the issue by moving where the version is set
and doing it in the more modern way: in the project() call.