Partly refactor tool progress reporting
Change AbstractTool::updateProgress to take a value, maximum pair rather
than a pre-normalized value (for current tools, this equates to the
active frame and last frame), rather than a normalized progress value,
and add a method for the caller to set the expected last frames.  This
saves a potentially expensive call to query the number of frames from
the video_input, when the GUI already knows the expected highest frame
number. (Note also that "last frame number" and "number of frames" are
not necessarily the same value, although we expect that in practice they
usually will be. However, since the progress being reported is the
current frame number, and not the number of frames processed, this
change also improves correctness in case the video skips frame numbers.)

This whole thing really needs to be refactored to use signals/slots and
to report the raw progress (right now, it is still being normalized in
AbstractTool), rather than sharing an int across threads with no
synchronization whatsoever. There are other changes that should be made
first, however, so punting on that for now...