Reverted location where tracks are updated
Previous changes move the location to where tracks are updated in the
CameraView to whenever the CameraView is updated for the frame changes.
This was an attemp to fix display issues, but it was not correct.
The track display data should only update when the underlying track
data updates.  That is, when tracks are loaded or computed in a tool.