Remove "depth" from ReconstructionData
The ReconstructionData class was ported from another project where
it was originally used to hold depth maps and cameras for reconstruction.
It was then adapted to also hold RGB images and cameras for coloration.
In TeleSculptor it has only ever been used for coloration, so this commit
replaces references to depth maps with images.