Fix build
Take immutable rather than mutable references for ReconstructionData and
vitalToVtkImage. These don't need the mutable reference (especially the
instances that took a smart pointer!), and trying to pass a reference
was causing a build error in MeshColoration::initializeDataList which
was passing an r-value, which cannot convert to a reference type.