Address clazy-detaching-member warnings
Tweak some code to avoid operating on mutable containers where not
necessary. Refactor a bunch of code to use qtGet.

This silences the clazy warnings, although some of the ones using qtGet
are silenced only because the warning would happen in qtGet.h rather
than a local source file. Also, all of the containers in question are
ones that most likely are not going to be shared, anyway. However, not
operating on a mutable container should at least save us needing to
check the reference count, and the qtGet calls are generally "tidier"
than before.