Reserve space in output vector
Tweak the code that converts a landmark map to a vector to first reserve
space in the output vector. This is likely slightly more efficient.

(There are other spots reported by clazy that might benefit from similar
changes, but none of them, including this one, are likely to be in
sufficiently "hot" paths to matter. This one, however, differs in that
the reserved size is likely to be moderately large, while the others are
likely operating on lists of single-digit size.)