Use the intersection point of camera dop with ground plane
If the depth estimation picker does not find a suitable point in the
scene along the camera direction of projection within a specific
tolerance, we find the intersection point of the camera vector with the
z=0 ground plane. This point is used as the ground control point in the
world view.