Fix issue where the gcp was picking along the camera path
When placing a point in the camera view, the point picker in the world
view was picks points along the camera direction vector. When there are
more than one cameras, there is an additional actor (path actor) in the
view. The point picker was picking points on this actor resulting in bad
depth estimation.