Implement GCP list view
Implement the list view, and backing model, for the ground control
points pane. The actually list is functioning and is mostly "done", but
the "details" part of the pane is not yet hooked up.

Also, fix copyright dates on GroundControlPointsView.{cpp,h}; these
files are new as of bd752a1ecdd1, but picked up incorrect copyright
years due to the header being copied from other files.