Show 'user registered' GCP's in list
Modify ground control point list to show an icon next to points for
which the user has specified the geodetic location. The icon is an SVG
which is rendered by Qt to color-match the user's color palette and in
several sizes to try to match the desired icon size of the list.
(Essentially, it tries to behave like a custom font glyph.) The code is
high-DPI aware and should behave correctly in the face of screen and
palette changes.