Add API to delete, reset GCP's
Add methods to GroundControlPointsHelper to delete or "reset" (remove
user-specified geodetic location and "go back" to the computed location)
ground control points.

Ground control points can already be deleted via the widget (i.e. during
interaction with a point in one of the graphical views). This adds the
ability to delete a GCP from outside the widget.