Improve seed handle lookup
Implement a new method in vtkMaptkSeedWidget to find the ID of a handle.
This allows us to improve looking up the VTK ID of a ground control
point (given the pointer to its handle) from O(n²) to O(n). Also, make a
similar tweak to vtkMaptkSeedWidget::HighlightActiveSeed which yields a
like reduction in complexity order.