Add signal to notify on batch changes to GCP's
Add a new signal (currently not emitted by anyone) to notify when the
data of "all" ground control points has changed, but the set of ground
control points has not. Hook up necessary observers. This will be useful
once we implement computed geodetic locations to notify when these have
been recomputed, as emitting pointsReloaded for such cases is
unnecessarily destructive. (In particular, pointsReloaded triggers a
model reset of the GCP list, which is not needed when only the data has
changed, and has unfortunate side effects such as causing the selection
to be lost. In fact, the model doesn't even need to be notified of data
changes; all we need to do is make sure the "details" display for the
current point, if any, is refreshed.)