Make mesh coloratation much faster by avoiding exception handling
When switching to using kwiver data structures we used to built-in
bounds checking on image access to catch out of bounds pixels and throw
an exception.  It turns out throwing and catching the exception in the
inner loops is too expensive because it allocates the exception message
each time.  It's far more efficient to pre-check the bounds.  The exception
is still handled in the code, but should never happen now.