Fix GCP elevation range
Fix the minimum and maximum values for the elevation field for the
currently selected ground control point. Somehow, the minimum was being
specified, but not the maximum (almost certainly this happened by
accident and/or oversight). The new range is on the order of "the
deepest we have ever drilled" to "the edge of space", which should be
sufficient for just about any terrestrial purpose.